#46: address some floating point determinism issues

+++ src/gradient.cpp
@@ -21,16 +21,16 @@
 
 // https://martin.ankerl.com/2012/01/25/optimized-approximative-pow-in-c-and-cpp/
 // an approximation to pow
-double fastPrecisePow(double a, double b) {
+float fastPrecisePow(float a, float b) {
   // calculate approximation with fraction of the exponent
-  int e = (int)b;
+  int e = static_cast<int>(b);
   union {
     double d;
     int x[2];
-  } u = {a};
-  u.x[1] = (int)((b - e) * (u.x[1] - 1072632447) + 1072632447);
+  } u = { a };
+  u.x[1] = static_cast<int>((b - e) * (u.x[1] - 1072632447) + 1072632447);
   u.x[0] = 0;
-
+  
   // exponentiation by squaring with the exponent's integer part
   // double r = u.d makes everything much slower, not sure why
   double r = 1.0;
@@ -41,26 +41,26 @@ double fastPrecisePow(double a, double b) {
     a *= a;
     e >>= 1;
   }
-
-  return r * u.d;
+  
+  return static_cast<float>(r * u.d);
 }
 
 // UMAP implementation code
-template <double (*powfun)(double, double)>
-base_umap_gradient<powfun>::base_umap_gradient(const double a, const double b,
-                                               const double gamma)
+template <float (*powfun)(float, float)>
+base_umap_gradient<powfun>::base_umap_gradient(const float a, const float b,
+                                               const float gamma)
     : a(a), b(b), a_b_m2(-2.0 * a * b), gamma_b_2(2.0 * gamma * b) {}
 
-template <double (*powfun)(double, double)>
-const double
-base_umap_gradient<powfun>::grad_attr(const double dist_squared) const {
-  const double pd2b = powfun(dist_squared, b);
+template <float (*powfun)(float, float)>
+const float
+base_umap_gradient<powfun>::grad_attr(const float dist_squared) const {
+  const float pd2b = powfun(dist_squared, b);
   return (a_b_m2 * pd2b) / (dist_squared * (a * pd2b + 1.0));
 }
 
-template <double (*powfun)(double, double)>
-const double
-base_umap_gradient<powfun>::grad_rep(const double dist_squared) const {
+template <float (*powfun)(float, float)>
+const float
+base_umap_gradient<powfun>::grad_rep(const float dist_squared) const {
   return gamma_b_2 /
          ((0.001 + dist_squared) * (a * powfun(dist_squared, b) + 1.0));
 }
@@ -74,23 +74,23 @@ template class base_umap_gradient<fastPrecisePow>;
 
 tumap_gradient::tumap_gradient() {}
 
-const double tumap_gradient::grad_attr(const double dist_squared) const {
+const float tumap_gradient::grad_attr(const float dist_squared) const {
   return -2.0 / (dist_squared + 1.0);
 }
 
-const double tumap_gradient::grad_rep(const double dist_squared) const {
+const float tumap_gradient::grad_rep(const float dist_squared) const {
   return 2.0 / ((0.001 + dist_squared) * (dist_squared + 1.0));
 }
 
 // LargeVis
 
-largevis_gradient::largevis_gradient(const double gamma)
+largevis_gradient::largevis_gradient(const float gamma)
     : gamma_2(gamma * 2.0) {}
 
-const double largevis_gradient::grad_attr(const double dist_squared) const {
+const float largevis_gradient::grad_attr(const float dist_squared) const {
   return -2.0 / (dist_squared + 1.0);
 }
 
-const double largevis_gradient::grad_rep(const double dist_squared) const {
+const float largevis_gradient::grad_rep(const float dist_squared) const {
   return gamma_2 / ((0.1 + dist_squared) * (dist_squared + 1.0));
 }
