+++ src/gradient.h
@@ -23,26 +23,26 @@
 #include <cmath>
 
 // Class templated on the powfun function as suggested by Aaron Lun
-template <double (*powfun)(double, double)> class base_umap_gradient {
+template <float (*powfun)(float, float)> class base_umap_gradient {
 public:
-  base_umap_gradient(const double a, const double b, const double gamma);
-  const double grad_attr(const double dist_squared) const;
-  const double grad_rep(const double dist_squared) const;
-  static const constexpr double clamp_hi = 4.0;
-  static const constexpr double clamp_lo = -4.0;
+  base_umap_gradient(const float a, const float b, const float gamma);
+  const float grad_attr(const float dist_squared) const;
+  const float grad_rep(const float dist_squared) const;
+  static const constexpr float clamp_hi = 4.0;
+  static const constexpr float clamp_lo = -4.0;
 
 private:
-  const double a;
-  const double b;
-  const double a_b_m2;
-  const double gamma_b_2;
+  const float a;
+  const float b;
+  const float a_b_m2;
+  const float gamma_b_2;
 };
 
 // UMAP
 typedef base_umap_gradient<std::pow> umap_gradient;
 
 // apUMAP: UMAP with an approximate power calculation
-double fastPrecisePow(double, double);
+float fastPrecisePow(float, float);
 typedef base_umap_gradient<fastPrecisePow> apumap_gradient;
 
 // t-UMAP: the UMAP function with a = 1, and b = 1, which results in the Cauchy
@@ -54,22 +54,22 @@ typedef base_umap_gradient<fastPrecisePow> apumap_gradient;
 class tumap_gradient {
 public:
   tumap_gradient();
-  const double grad_attr(const double dist_squared) const;
-  const double grad_rep(const double dist_squared) const;
-  static const constexpr double clamp_hi = 4.0;
-  static const constexpr double clamp_lo = -4.0;
+  const float grad_attr(const float dist_squared) const;
+  const float grad_rep(const float dist_squared) const;
+  static const constexpr float clamp_hi = 4.0;
+  static const constexpr float clamp_lo = -4.0;
 };
 
 class largevis_gradient {
 public:
-  largevis_gradient(const double gamma);
-  const double grad_attr(const double dist_squared) const;
-  const double grad_rep(const double dist_squared) const;
-  static const constexpr double clamp_hi = 5.0;
-  static const constexpr double clamp_lo = -5.0;
+  largevis_gradient(const float gamma);
+  const float grad_attr(const float dist_squared) const;
+  const float grad_rep(const float dist_squared) const;
+  static const constexpr float clamp_hi = 5.0;
+  static const constexpr float clamp_lo = -5.0;
 
 private:
-  const double gamma_2;
+  const float gamma_2;
 };
 
 #endif // UWOT_GRADIENT_H
