+++ src/sampler.cpp
@@ -19,8 +19,8 @@
 
 #include "sampler.h"
 
-Sampler::Sampler(const std::vector<double> &epochs_per_sample,
-                 const double negative_sample_rate)
+Sampler::Sampler(const std::vector<float> &epochs_per_sample,
+                 const float negative_sample_rate)
     :
 
       epochs_per_sample(epochs_per_sample),
@@ -28,7 +28,7 @@ Sampler::Sampler(const std::vector<double> &epochs_per_sample,
       epochs_per_negative_sample(epochs_per_sample.size()),
       epoch_of_next_negative_sample(epochs_per_sample.size()) {
   const std::size_t esz = epochs_per_sample.size();
-  const double nsr = 1.0 / negative_sample_rate;
+  const float nsr = 1.0 / negative_sample_rate;
   for (std::size_t i = 0; i < esz; i++) {
     epochs_per_negative_sample[i] = epochs_per_sample[i] * nsr;
     epoch_of_next_negative_sample[i] = epochs_per_negative_sample[i];
